Rename Line -> Chunk, Keyword -> Kw
I wanted to use Line in places that weren't really a line,
so figure I should make it a "Chunk" instead, which just means
"display this as a block".

I disliked how Keyword was `<Keyword kw="class" />`, which
differed from everything else. So I changed it to be
`<Keyword>class</Keyword>`, but that's kind of verbose,
so I shortened `Keyword` to `Kw`.