Remove ES specific things... it's Ruby code
Trying to make it do the right thing for ES and Ruby was resulting in
either doing both badly or needing to turn this into a transpiler, which
isn't its goal. So, embrace the fact that it's a Ruby AST, the ES syntax
is just for decoration, not expected to be actually correct.